+++ cpp_source/utils/misc.h
@@ -20,7 +20,7 @@ std::string str(printable *obj);
 std::string str(printable &obj);
 std::string str(int num);
 std::string str(ulong num);
-std::string str(double num);
+std::string str(float num);
 
 std::ostream &operator<<(std::ostream &output, printable &p);
 std::ostream &operator<<(std::ostream &output, printable *p);
